requirements.txt
httpx
pandas
python-dotenv
anthropic
